Extract more vector routines to vector module
